+++ helloworld-blockchain-netcore/src/main/java/com/xingkaichun/helloworldblockchain/netcore/dto/netserver/request/QueryBlockDtoByBlockHeightRequest.java
@@ -1,27 +1,26 @@
 package com.xingkaichun.helloworldblockchain.netcore.dto.netserver.request;
 
-import java.math.BigInteger;
-
 /**
  *
  * @author 邢开春 微信HelloworldBlockchain 邮箱xingkaichun@qq.com
  */
 public class QueryBlockDtoByBlockHeightRequest {
 
-    private BigInteger blockHeight;
+    private Long blockHeight;
 
 
 
 
     //region get set
 
-    public BigInteger getBlockHeight() {
+    public Long getBlockHeight() {
         return blockHeight;
     }
 
-    public void setBlockHeight(BigInteger blockHeight) {
+    public void setBlockHeight(Long blockHeight) {
         this.blockHeight = blockHeight;
     }
+
 
     //endregion
 }
